refactor(repo-cli): close canonical slice factory v1
