+++ .changeset/canonical-slice-factory-v1-closure.md
@@ -1,0 +1,4 @@
+---
+---
+
+Close the canonical slice factory V1 initiative packet and split architecture operation-plan execution helpers without changing the public command grammar or operation-plan schema.
